requirements.txt
Adafruit-Blinka==6.4.1
adafruit-circuitpython-dht==3.5.8
Adafruit-PlatformDetect==3.4.1
Adafruit-PureIO==1.1.8
arrow==1.0.3
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
MarkupSafe==1.1.1
numpy==1.20.2
pandas==1.2.3
pyftdi==0.52.9
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
pyusb==1.1.1
rpi-ws281x==4.2.6
RPi.GPIO==0.7.0
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sysv-ipc==1.1.0
threadpoolctl==2.1.0
uWSGI==2.0.19.1
Werkzeug==1.0.1
